Rename llm package
