+++ external/open_ai/request.go
@@ -1,4 +1,4 @@
-package llm
+package open_ai
 
 import (
 	"context"
