+++ external/open_ai/test.go
@@ -1,4 +1,4 @@
-package llm
+package open_ai
 
 import (
 	"context"
