+++ internal/service/solution/generate.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"github.com/sashabaranov/go-openai"
 	"log"
-	"neko-acm/external/llm"
+	"neko-acm/external/open_ai"
 	"neko-acm/internal/model"
 	"neko-acm/prompt"
 	"neko-acm/utils"
@@ -23,12 +23,12 @@ func Generate(si model.SolutionInstruction) (model.Solution, error) {
 	log.Println("请求生成题解：" + instruction)
 
 	// 组合Prompt
-	sysMsg := llm.NewSysMsg(prompt.SolutionGenerate)
-	userMsg := llm.NewUserMsg(instruction)
+	sysMsg := open_ai.NewSysMsg(prompt.SolutionGenerate)
+	userMsg := open_ai.NewUserMsg(instruction)
 	msgs := []openai.ChatCompletionMessage{sysMsg, userMsg}
 
 	// 请求模型
-	resp, err := llm.RequestMessages(msgs)
+	resp, err := open_ai.RequestMessages(msgs)
 	if err != nil {
 		return model.Solution{}, errors.New("请求模型失败！")
 	}
