+++ internal/service/testcase/generate.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"github.com/sashabaranov/go-openai"
 	"log"
-	"neko-acm/external/llm"
+	"neko-acm/external/open_ai"
 	"neko-acm/internal/model"
 	"neko-acm/prompt"
 	"neko-acm/utils"
@@ -23,12 +23,12 @@ func Generate(ti model.TestcaseInstruction) (model.Testcase, error) {
 	log.Println("请求生成测试用例：" + instruction)
 
 	// 组合Prompt
-	sysMsg := llm.NewSysMsg(prompt.TestcaseGenerate)
-	userMsg := llm.NewUserMsg(instruction)
+	sysMsg := open_ai.NewSysMsg(prompt.TestcaseGenerate)
+	userMsg := open_ai.NewUserMsg(instruction)
 	msgs := []openai.ChatCompletionMessage{sysMsg, userMsg}
 
 	// 请求模型
-	resp, err := llm.RequestMessages(msgs)
+	resp, err := open_ai.RequestMessages(msgs)
 	if err != nil {
 		return model.Testcase{}, errors.New("请求模型失败！")
 	}
